Delete is now complete, add the id at SampleData and getId/setId methods

+++ app/src/main/java/com/example/nicola/paint/DBAdapter.java
@@ -53,7 +53,7 @@ public class DBAdapter {
 
 
     // elimina record
-    public boolean deleteRecord(int sessionID)
+    public boolean deleteRecord(long sessionID)
     {
         return database.delete(DBOpenHelper.TABLE, DBOpenHelper.ID + "=" + sessionID, null) > 0;
     }
